feat: new ui for watch only mode selector
