+++ frontend/pages/components/topbar/Pill.tsx
@@ -1,8 +1,4 @@
 import { shortAddress } from "@/utils";
-import { ChevronDownIcon, ChevronLeftIcon } from "@radix-ui/react-icons";
-import { useAppSelector } from "@redux/Store";
-import WatchOnlyRecords from "./WatchOnlyRecords";
-import { useState } from "react";
 
 interface PillProps {
   text: string;
@@ -12,34 +8,14 @@ interface PillProps {
 }
 
 export default function Pill({ text, start, end, icon }: PillProps) {
-  const { watchOnlyMode } = useAppSelector((state) => state.auth);
-  const [historicalOpen, setHistoricalOpen] = useState(false);
-
   return (
     <div className="relative w-[16rem]">
       <div className="px-3 py-1 rounded-full bg-GrayColor/50">
         <div className="flex items-center justify-center w-full gap-2 whitespace-nowrap">
           <img src={icon} alt="icon" className="w-5" />
           {shortAddress(text, start, end)}
-          {watchOnlyMode && (
-            <>
-              {historicalOpen ? <ChevronDownIcon className="w-4 h-4 cursor-pointer" onClick={onOpenChange} /> : null}
-              {!historicalOpen ? <ChevronLeftIcon className="w-4 h-4 cursor-pointer" onClick={onOpenChange} /> : null}
-            </>
-          )}
         </div>
       </div>
-
-      {watchOnlyMode && historicalOpen && (
-        <>
-          <div className="fixed inset-0 z-10" onClick={onOpenChange}></div>
-          <WatchOnlyRecords start={start} end={end} />
-        </>
-      )}
     </div>
   );
-
-  function onOpenChange() {
-    setHistoricalOpen((prev) => !prev);
-  }
 }
